+++ extensions/all/misc/screenshot/remove-screenshot-restriction/src/main/java/app/revanced/extension/all/misc/screenshot/removerestriction/RemoveScreenshotRestrictionPatch.java
@@ -1,4 +1,4 @@
-package app.revanced.extension.all.screenshot.removerestriction;
+package app.revanced.extension.all.misc.screenshot.removerestriction;
 
 import android.view.Window;
 import android.view.WindowManager;