+++ extensions/shared/library/src/main/java/app/revanced/extension/shared/spoof/requests/StreamingDataRequest.java
@@ -204,7 +204,7 @@ public class StreamingDataRequest {
                     // but empty response body does.
                     if (connection.getContentLength() == 0) {
                         if (BaseSettings.DEBUG.get() && BaseSettings.DEBUG_TOAST_ON_ERROR.get()) {
-                            Utils.showToastShort("Ignoring empty spoof stream client: " + clientType);
+                            Utils.showToastShort("Debug: Ignoring empty spoof stream client " + clientType);
                         }
                     } else {
                         try (InputStream inputStream = new BufferedInputStream(connection.getInputStream());