+++ patches/src/main/kotlin/app/revanced/patches/twitter/misc/hook/HookPatch.kt
@@ -13,8 +13,8 @@ fun hookPatch(
 
     compatibleWith(
         "com.twitter.android"(
-            // 10.85+ uses Pairip and requires additional changes to work.
-            "10.84.0-release.0",
+            // Only v10.85 uses Pairip and requires additional changes to work.
+            "10.86.0-release.0",
             // Confirmed to not show reply ads. Slightly newer versions may also work.
             "10.60.0-release.0",
             "10.48.0-release.0"