+++ patches/src/main/kotlin/app/revanced/patches/twitter/misc/links/ChangeLinkSharingDomainPatch.kt
@@ -39,7 +39,7 @@ val changeLinkSharingDomainPatch = bytecodePatch(
 
     compatibleWith(
         "com.twitter.android"(
-            "10.84.0-release.0",
+            "10.86.0-release.0",
             "10.60.0-release.0",
             "10.48.0-release.0"
         )