+++ patches/src/main/kotlin/app/revanced/patches/youtube/interaction/seekbar/EnableSeekbarTappingPatch.kt
@@ -10,9 +10,14 @@ import app.revanced.patches.shared.misc.settings.preference.SwitchPreference
 import app.revanced.patches.youtube.misc.extension.sharedExtensionPatch
 import app.revanced.patches.youtube.misc.settings.PreferenceScreen
 import app.revanced.patches.youtube.misc.settings.settingsPatch
+import app.revanced.util.findFreeRegister
+import app.revanced.util.indexOfFirstInstructionOrThrow
+import com.android.tools.smali.dexlib2.Opcode
+import com.android.tools.smali.dexlib2.iface.instruction.FiveRegisterInstruction
 import com.android.tools.smali.dexlib2.iface.instruction.ReferenceInstruction
-import com.android.tools.smali.dexlib2.iface.instruction.formats.Instruction35c
 import com.android.tools.smali.dexlib2.iface.reference.MethodReference
+
+private const val EXTENSION_CLASS_DESCRIPTOR = "Lapp/revanced/extension/youtube/patches/SeekbarTappingPatch;"
 
 val enableSeekbarTappingPatch = bytecodePatch(
     description = "Adds an option to enable tap to seek on the seekbar of the video player.",
@@ -31,39 +36,37 @@ val enableSeekbarTappingPatch = bytecodePatch(
         )
 
         // Find the required methods to tap the seekbar.
-        val patternMatch = onTouchEventHandlerFingerprint.patternMatch!!
+        val seekbarTappingMethods = onTouchEventHandlerFingerprint.let {
+            fun getMethodReference(index: Int) = it.method.getInstruction<ReferenceInstruction>(index)
+                .reference as MethodReference
 
-        fun getReference(index: Int) = onTouchEventHandlerFingerprint.method.getInstruction<ReferenceInstruction>(index)
-            .reference as MethodReference
-
-        val seekbarTappingMethods = buildMap {
-            put("N", getReference(patternMatch.startIndex))
-            put("O", getReference(patternMatch.endIndex))
+            listOf(
+                getMethodReference(it.patternMatch!!.startIndex),
+                getMethodReference(it.patternMatch!!.endIndex)
+            )
         }
 
-        val insertIndex = seekbarTappingFingerprint.patternMatch!!.endIndex - 1
-
         seekbarTappingFingerprint.method.apply {
-            val thisInstanceRegister = getInstruction<Instruction35c>(insertIndex - 1).registerC
+            val pointIndex = indexOfNewPointInstruction(this)
+            val invokeIndex = indexOfFirstInstructionOrThrow(pointIndex, Opcode.INVOKE_VIRTUAL)
+            val insertIndex = invokeIndex + 1
 
-            val freeRegister = 0
-            val xAxisRegister = 2
+            val thisInstanceRegister = getInstruction<FiveRegisterInstruction>(invokeIndex).registerC
+            val xAxisRegister = this.getInstruction<FiveRegisterInstruction>(pointIndex).registerD
+            val freeRegister = findFreeRegister(insertIndex, thisInstanceRegister, xAxisRegister)
 
-            val oMethod = seekbarTappingMethods["O"]!!
-            val nMethod = seekbarTappingMethods["N"]!!
-
-            fun MethodReference.toInvokeInstructionString() =
-                "invoke-virtual { v$thisInstanceRegister, v$xAxisRegister }, $this"
+            val oMethod = seekbarTappingMethods[0]
+            val nMethod = seekbarTappingMethods[1]
 
             addInstructionsWithLabels(
                 insertIndex,
                 """
-                        invoke-static { }, Lapp/revanced/extension/youtube/patches/SeekbarTappingPatch;->seekbarTappingEnabled()Z
-                        move-result v$freeRegister
-                        if-eqz v$freeRegister, :disabled
-                        ${oMethod.toInvokeInstructionString()}
-                        ${nMethod.toInvokeInstructionString()}
-                    """,
+                    invoke-static { }, $EXTENSION_CLASS_DESCRIPTOR->seekbarTappingEnabled()Z
+                    move-result v$freeRegister
+                    if-eqz v$freeRegister, :disabled
+                    invoke-virtual { v$thisInstanceRegister, v$xAxisRegister }, $oMethod
+                    invoke-virtual { v$thisInstanceRegister, v$xAxisRegister }, $nMethod
+                """,
                 ExternalLabel("disabled", getInstruction(insertIndex)),
             )
         }