+++ patches/src/main/kotlin/app/revanced/patches/youtube/video/speed/PlaybackSpeedPatch.kt
@@ -34,6 +34,7 @@ val playbackSpeedPatch = bytecodePatch(
             "19.43.41",
             "19.47.53",
             "20.07.39",
+            "20.12.46",
         )
     )